Move label-rewriting to its own file.
It's pretty independent of the rest of the logic, with a nice
narrow interface, and is sort of a side detail vs. the main point.
